+++ src/main/java/ua/yakovenko/model/dao/impl/UserJdbcDao.java
@@ -113,12 +113,12 @@ public class UserJdbcDao implements UserDao {
     }
 
     @Override
-    public void delete(int id) {
+    public void delete(Long id) {
         try (PreparedStatement ps =
                      connection.prepareStatement(
                             QUERY_USER_DELETE_BY_ID)
         ) {
-            ps.setInt(1, id);
+            ps.setLong(1, id);
 
             ps.executeUpdate();
         } catch (SQLException e) {
